[RFE] Improve TS2.0 report parser method to adapt frequent TS2.0 change
Comparing the previous methods with the improved ones, the improved one
is more powerful and flexible. It can support the TS2.0 workflow change
well. See the details below:
1. Previous parser method consider the TS2.0 steps are fixed then do the
according parser. It means the the parser cannot catch the failure
correctly based on the TS2.0 workflow changes. I always need to fix it.
2. The improved method get all steps and status as one dictory.
It will check the steps_status_map keys to check the TS2.0 status.
Once we keep the name of steps are stable, CI_3 can always work well